move HTML serving endpoints into separate module
