+++ src/api/src/lib.rs
@@ -25,7 +25,7 @@ pub mod events;
 pub mod fed_cm;
 pub mod generic;
 pub mod groups;
-pub mod i18n;
+pub mod html;
 pub mod oidc;
 pub mod openapi;
 pub mod roles;
